add output hiddens
