+++ vllm/model_executor/layers/logits_processor.py
@@ -1,6 +1,6 @@
 """A layer that compute logits from hidden_stats."""
 import inspect
-from typing import Optional
+from typing import Optional, Union
 
 import torch
 import torch.nn as nn
@@ -70,12 +70,15 @@ class LogitsProcessor(nn.Module):
         return logits
 
     def _get_logits(self, hidden_states: torch.Tensor,
-                    lm_head: VocabParallelEmbedding,
+                    lm_head: Union[VocabParallelEmbedding, nn.Linear],
                     embedding_bias: Optional[torch.Tensor]) -> torch.Tensor:
         # Get the logits for the next tokens.
-        logits = lm_head.linear_method.apply(lm_head,
-                                             hidden_states,
-                                             bias=embedding_bias)
+        if isinstance(lm_head, nn.Linear):
+            logits = lm_head(hidden_states)
+        else:
+            logits = lm_head.linear_method.apply(lm_head,
+                                                hidden_states,
+                                                bias=embedding_bias)
         logits = tensor_model_parallel_gather(logits)
         # Remove paddings in vocab (if any).
         if logits is not None:
